game structure build, should work in theory but for testing window stuff i needed. coming back to that

+++ inc/parser_cub.h
@@ -24,7 +24,7 @@ typedef struct s_map
 {
 	int		x_len;
 	int		y_len;
-	int		player_pos[2];
+	double		player_pos[2];
 	int		player_dir;
 	char	*map_storage;
 	char	**map;
